// app/server/applicationsDB.server/applications/pcsx2/types.ts
export type Pcsx2ButtonId =
  | "Up"
  | "Left"
  | "Right"
  | "Down"
  | "LUp"
  | "LLeft"
  | "LRight"
  | "LDown"
  | "Cross"
  | "Circle"
  | "Square"
  | "Triangle"
  | "L2"
  | "R2"
  | "L1"
  | "R1"
  | "Select"
  | "Start"
  | "L3"
  | "R3"
  | "RUp"
  | "RLeft"
  | "RRight"
  | "RDown"
  | "Analog";